Update formatter documentation
